Moved the code relating to the text complexity to text_complexity.py
